+++ src/main/java/chess/domain/converter/PieceConverter.java
@@ -1,4 +1,4 @@
-package chess.converter;
+package chess.domain.converter;
 
 import chess.domain.piece.*;
 import chess.domain.piece.attribute.Color;
